rename seller manager directives to salespeople
