+++ src/app/management/management-routing.module.ts
@@ -14,8 +14,8 @@ import { ProductManagerAccessResolver } from './routes/products/product-manager.
 import { ProductManagerComponent } from './routes/products/product-manager.component';
 import { SellManagerAccessResolver } from './routes/sales/sell-manager.access-resolver';
 import { SellManagerComponent } from './routes/sales/sell-manager.component';
-import { SellerManagerAccessResolver } from './routes/sellers/seller-manager.access-resolver';
-import { SellerManagerComponent } from './routes/sellers/seller-manager.component';
+import { SalespersonManagerAccessResolver } from './routes/salespeople/salesperson-manager.access-resolver';
+import { SalespersonManagerComponent } from './routes/salespeople/salesperson-manager.component';
 import { UserManagerAccessResolver } from './routes/users/user-manager.access-resolver';
 import { UserManagerComponent } from './routes/users/user-manager.component';
 
@@ -44,9 +44,9 @@ export const MANAGEMENT_CHILD_ROUTES: ManagementChildRoute[] = [
     resolve: { access: CustomerManagerAccessResolver }
   },
   {
-    path: 'sellers', component: SellerManagerComponent,
+    path: 'salespeople', component: SalespersonManagerComponent,
     data: { matIcon: 'work_outline', title: 'Vendedores' },
-    resolve: { access: SellerManagerAccessResolver }
+    resolve: { access: SalespersonManagerAccessResolver }
   },
   {
     path: 'users', component: UserManagerComponent,
