+++ src/app/management/routes/sales/form-dialog/sell-manager-form-dialog.component.ts
@@ -102,7 +102,7 @@ export class SellManagerFormDialogComponent
     this.sellDetails$ = this.service.sellDetails$.pipe(tap(details => { this.sellDetails = details; }));
 
     this.sellTypes$ = this.service.getAllSellTypes();
-    this.sellers$ = this.service.getAllSellers();
+    this.sellers$ = this.service.getAllSalespeople();
     this.customers$ = this.service.getAllCustomers();
 
     this.sellSubtotalValue$ = this.service.sellSubtotalValue$.pipe();
