+++ src/app/management/routes/sales/form-dialog/sell-manager-form.service.ts
@@ -34,7 +34,7 @@ export class SellManagerFormService
     @Inject(API_SERVICE_INJECTION_TOKENS.salesCrud) protected dataService: CompositeEntityDataApiIService<Sell, SellDetail>,
     @Inject(API_SERVICE_INJECTION_TOKENS.productsCrud) protected productDataService: EntityDataApiIService<Product>,
     @Inject(API_SERVICE_INJECTION_TOKENS.customersCrud) protected customersDataService: EntityDataApiIService<Customer>,
-    @Inject(API_SERVICE_INJECTION_TOKENS.sellersCrud) protected sellerDataService: EntityDataApiIService<Salesperson>,
+    @Inject(API_SERVICE_INJECTION_TOKENS.salespeopleCrud) protected salespeopleDataService: EntityDataApiIService<Salesperson>,
     @Inject(API_SERVICE_INJECTION_TOKENS.shared) protected sharedDataService: SharedDataApiIService,
   ) {
     super();
@@ -58,8 +58,8 @@ export class SellManagerFormService
     return this.sharedDataService.readAllSellTypes();
   }
 
-  public getAllSellers(): Observable<Salesperson[]> {
-    return this.sellerDataService.readAll();
+  public getAllSalespeople(): Observable<Salesperson[]> {
+    return this.salespeopleDataService.readAll();
   }
 
   public getAllCustomers(): Observable<Customer[]> {
