+++ src/app/management/routes/salespeople/form-dialog/seller-manager-form-dialog.component.html
@@ -1,0 +1,32 @@
+<!--
+ Copyright (c) 2020 Benjamin
+
+ This software is released under the MIT License.
+ https://opensource.org/licenses/MIT
+-->
+
+<div class="dialog">
+  <h2 mat-dialog-title>{{ dialogTitle }}</h2>
+  <centered-mat-spinner *ngIf="saving$ | async">
+  </centered-mat-spinner>
+
+  <div [hidden]="saving$ | async">
+    <form [formGroup]="formGroup" (submit)="onSubmit()">
+
+      <app-person-form #personForm>
+      </app-person-form>
+
+      <div class="actions">
+        <button mat-button mat-raised-button type="submit"
+          color="primary"
+          [disabled]="formGroup.invalid">
+          Aceptar
+        </button>
+        <button mat-button mat-raised-button type="button"
+          (click)="onCancel()">
+          Cancelar
+        </button>
+      </div>
+    </form>
+  </div>
+</div>
